+++ cocotools/src/converters.rs
@@ -1,3 +1,3 @@
 //! Module containing utilities to convert data formats.
 
-pub mod masks;
+pub mod mask;
