+++ cocotools/src/converters/mask.rs
